Moved verified attributes actions logic to a separate hook
