+++ src/components/shared/AgreementDetails/components/AgreementAttributesListSections/AgreementVerifiedAttributesSection.tsx
@@ -1,9 +1,7 @@
 import React from 'react'
-import { useJwt } from '@/hooks/useJwt'
 import { useCurrentRoute } from '@/router'
 import { useTranslation } from 'react-i18next'
 import { useAgreementDetailsContext } from '../../AgreementDetailsContext'
-import { AttributeMutations } from '@/api/attribute'
 import {
   SectionContainer,
   AttributeGroupContainer,
@@ -17,77 +15,26 @@ import {
   isAttributeRevoked,
 } from '@/utils/attribute.utils'
 import type { ProviderOrConsumer } from '@/types/common.types'
+import { useAgreementGetVerifiedAttributesActions } from '../../hooks/useAgreementGetVerifiedAttributesActions'
 
 export const AgreementVerifiedAttributesSection: React.FC = () => {
-  const { t } = useTranslation('agreement', { keyPrefix: 'read.attributes' })
   const { t: tAttribute } = useTranslation('attribute')
   const { mode } = useCurrentRoute()
-  const { isAdmin } = useJwt()
 
-  const { agreement, eserviceAttributes, partyAttributes, isAgreementEServiceMine } =
-    useAgreementDetailsContext()
-
-  const { mutate: verifyAttribute } = AttributeMutations.useVerifyPartyAttribute()
-  const { mutate: revokeAttibute } = AttributeMutations.useRevokeVerifiedPartyAttribute()
+  const { eserviceAttributes, partyAttributes } = useAgreementDetailsContext()
 
   const providerOrConsumer = mode as ProviderOrConsumer
 
   const verifiedAttributeGroups = eserviceAttributes?.verified ?? []
   const ownedVerifiedAttributes = partyAttributes?.verified ?? []
 
-  const handleVerifyAttribute = (attributeId: string) => {
-    if (!agreement?.consumer.id) return
-    verifyAttribute({
-      partyId: agreement.consumer.id,
-      id: attributeId,
-      renewal: 'AUTOMATIC_RENEWAL',
-    })
-  }
-
-  const handleRevokeAttribute = (attributeId: string) => {
-    if (!agreement?.consumer.id) return
-    revokeAttibute({
-      partyId: agreement.consumer.id,
-      attributeId,
-    })
-  }
-
-  const isVerifiedAttributeRevoked = (attributeId: string) => {
-    const attribute = ownedVerifiedAttributes.find((a) => a.id === attributeId)
-    return attribute && isAttributeRevoked('verified', attribute)
-  }
-
-  const getAttributeActions = (attributeId: string) => {
-    // The user can certify verified attributes in this view only if it is a provider...
-    if (!agreement || mode === 'consumer' || !isAdmin) return []
-    // ... only if the e-service does not belong to itself
-    if (isAgreementEServiceMine) return []
-    // ... and only if the agreement is active, pending or suspended
-    if (!['ACTIVE', 'PENDING', 'SUSPENDED'].includes(agreement.state)) return []
-
-    const isOwned = isAttributeOwned('verified', attributeId, ownedVerifiedAttributes)
-    const isRevoked = isVerifiedAttributeRevoked(attributeId)
-
-    const attributeActions: React.ComponentProps<typeof AttributeContainer>['actions'] = [
-      {
-        label: isOwned ? t('verified.actions.update') : t('verified.actions.verify'),
-        action: handleVerifyAttribute,
-      },
-    ]
-
-    if (isOwned && !isRevoked) {
-      attributeActions.push({
-        label: t('verified.actions.revoke'),
-        action: handleRevokeAttribute,
-        color: 'error',
-      })
-    }
-
-    return attributeActions
-  }
+  const getAttributeActions = useAgreementGetVerifiedAttributesActions()
 
   const getChipLabel = (attributeId: string) => {
-    if (mode === 'consumer' && isVerifiedAttributeRevoked(attributeId))
+    const attribute = ownedVerifiedAttributes.find((a) => a.id === attributeId)
+    if (!attribute) return
+
+    if (mode === 'consumer' && isAttributeRevoked('verified', attribute))
       return tAttribute('group.manage.revokedByParty')
   }
